Uploads are temporary by default

+++ apps/api/src/tus/cleanup.ts
@@ -2,7 +2,7 @@ import logger from "../services/log";
 import TusUploadModel, { TusUpload } from "./model";
 import { removeTusFiles } from "./utils";
 
-export async function Cleanup() {
+export async function cleanupTUSUploads() {
     logger.info({}, "Starting the tus uploads cleanup job");
 
     const now = new Date();
@@ -10,10 +10,23 @@ export async function Cleanup() {
         expiresAt: { $lt: now },
     }).lean()) as unknown as TusUpload[];
 
+    if (expiredUploads.length === 0) {
+        logger.info("No expired tus uploads found to cleanup");
+        return;
+    }
+
+    logger.info(
+        { count: expiredUploads.length },
+        "Found expired tus uploads to cleanup",
+    );
+
     for (const expiredUpload of expiredUploads) {
         removeTusFiles(expiredUpload.tempFilePath);
         await TusUploadModel.deleteOne({ _id: (expiredUpload as any)._id });
     }
 
-    logger.info({}, "Ending the tus uploads cleanup job");
+    logger.info(
+        { count: expiredUploads.length },
+        "Cleaned up expired tus uploads",
+    );
 }
